python/requirements.txt
# Mem0 REST API Server Dependencies

# Core
mem0ai>=0.1.0
requests>=2.31.0

# For .env file support
python-dotenv>=1.0.0

# Optional: For better JSON handling
pydantic>=2.0.0

# Local Model Server Dependencies
# (Uncomment or install these for local model support)
torch>=2.0.0
transformers>=4.40.0
pillow>=10.0.0
numpy>=1.24.0
sentencepiece>=0.2.0
protobuf>=4.0.0
huggingface_hub>=0.20.0
accelerate>=0.25.0
